Update color.py
change line581 for pytorch tensor

+++ imgaug/augmenters/color.py
@@ -578,7 +578,7 @@ class ChangeColorspace(meta.Augmenter):
                     img_to_cs = img_to_cs[:, :, np.newaxis]
                     img_to_cs = np.tile(img_to_cs, (1, 1, 3))
 
-                result[i] = blend.blend_alpha(img_to_cs, image, alpha, self.eps)
+                result[i] = blend.blend_alpha(img_to_cs.astype(np.float32), image, alpha, self.eps)
 
         return images
